Simplify synchronization and UART
Using an interrupt on the UART RX pin may not work on all platforms (it actually doesn't work with ESP32 Arduino core v3.0.0 Alpha 1).
This does away with the interrupt synchronization, and instead just looks at the received data in the UART buffer. A break signal just looks like an extra 0 byte.
There is a risk of improper synchronization, but the bytes are checked for zeros and the buffer is flushed if it looks wrong. Testing has shown it to work well.
This also removes the requirement for the DMX class to know the TX and RX pins, another simplification
Also moved _dmxBuffer.begin() outside of begin(), caller is required to begin the serial port with the proper settings. This makes it easier to ensure pins and such are set properly on any platform